server/requirements.txt
mcp
gspread
google-auth
